+++ src/plugins/valueHandler/basic/basicValueHandler.component.ts
@@ -51,7 +51,7 @@ export class BasicValueHandlerComponent<TValue = any> extends ValueHandlerBase<T
 
     //######################### constructor #########################
     constructor(@Inject(NG_SELECT_PLUGIN_INSTANCES) @Optional() ngSelectPlugins: NgSelectPluginInstances,
-                @Optional() pluginBus: PluginBus,
+                @Optional() pluginBus: PluginBus<TValue>,
                 pluginElement: ElementRef,
                 @Inject(VALUE_HANDLER_OPTIONS) @Optional() options?: BasicValueHandlerOptions)
     {